Code Clean Up + Minor Fixes/Changes
Returned Wyvern Ostrich flying mechanism to original state.

Replaced Fire Ogres dropping Heart of Fire with Gold Ingots.

Werewolves now howl when they transform into wolf form.

Cleaned up code by removing duplicate functions and unused code.

Fixed shark attacking players on boats when difficulty is below hard.

Fixed Wyvern Lair Tall Grass not spawning in Wyvern Lair.

Wraiths no longer attack players who are riding undead mounts unless they are attacked.

+++ src/main/java/drzhark/mocreatures/client/renderer/entity/MoCRenderDolphin.java
@@ -194,7 +194,7 @@ public class MoCRenderDolphin extends RenderLiving {
     {
     	if(entityDolphin.riddenByEntity != null && !entityDolphin.onGround)
     	{
-			GL11.glRotatef((float) entityDolphin.getRotationAmount(), -1, 0, 0);
+			GL11.glRotatef(entityDolphin.getRotationAmount(), -1, 0, 0);
     	}
     }
